merger of window and frame command classes
