+++ MySession/src/main/java/commands/ElementsOp.java
@@ -1,6 +1,5 @@
 package commands;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.TakesScreenshot;
